Modify Proposal component to be reusable
