+++ app/[eventSlug]/proposal.tsx
@@ -1,20 +1,13 @@
 "use client";
 
-import { useContext } from "react";
-import Link from "next/link";
-import { PencilIcon } from "@heroicons/react/24/outline";
-
-import { UserContext } from "@/app/context";
 import type { Guest } from "@/db/guests";
 import type { SessionProposal } from "@/db/sessionProposals";
 
-export function ViewProposal(props: {
+export function Proposal(props: {
   proposal: SessionProposal;
   guests: Guest[];
-  eventSlug: string;
 }) {
-  const { proposal, guests, eventSlug } = props;
-  const { user: currentUserId } = useContext(UserContext);
+  const { proposal, guests } = props;
   const displayDuration = (duration: number) => {
     if (duration === 30) {
       return "30 minutes";
@@ -22,13 +15,6 @@ export function ViewProposal(props: {
       const numHours = duration / 60;
       const hoursStr = numHours === 1 ? "hour" : "hours";
       return `${numHours} ${hoursStr}`;
-    }
-  };
-  const canEdit = () => {
-    if (proposal.hosts.length === 0) {
-      return true;
-    } else {
-      return currentUserId && proposal.hosts.includes(currentUserId);
     }
   };
   return (
@@ -45,15 +31,6 @@ export function ViewProposal(props: {
         <p className="text-sm text-gray-600 mb-4">
           Duration: {displayDuration(proposal.durationMinutes)}
         </p>
-      )}
-      {canEdit() && (
-        <Link
-          href={`/${eventSlug}/proposals/${proposal.id}/edit`}
-          className="text-rose-400 hover:text-rose-500 inline-flex items-center"
-        >
-          <PencilIcon className="h-4 w-4 mr-1" />
-          Edit
-        </Link>
       )}
     </>
   );
